+++ CHANGELOG.md
@@ -6,6 +6,50 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html),
 and is generated by [Changie](https://github.com/miniscruff/changie).
 
 
+## v1.8.2 - 2026-04-23
+### Added
+* `User-Agent` header now includes the library version (e.g. `aura-go-client/v1.8.0`) for easier identification in API server logs
+### Fixed
+* `TruncateString` now counts runes instead of bytes, preventing truncation mid-multibyte UTF-8 character
+* URL construction now trims stray leading/trailing slashes before joining, preventing double-slash in API endpoint paths
+* Added correctly-spelled `StatusRestoring` constant; retained misspelled `StatusRestroying` as a deprecated alias for backwards compatibility
+* `gDSSessionService.Get` and `Delete` now return an error for an empty session ID rather than sending a malformed request to the API
+* `GetInstanceHealth` no longer stacks two independent timeouts; it calls the internal `doFetchRawMetrics` directly so the context deadline is applied exactly once
+* `tenantService.List`, `Get`, and `GetMetrics` now check for a pre-cancelled context at entry, consistent with all other service methods
+* Removed redundant HTTP 200-status check in `FetchRawMetrics` — the API layer already converts any non-2xx response to an error before returning
+* Mock types moved from `internal/api` and `internal/httpClient` to `internal/testutil` — mocks no longer compile into production builds
+* Removed `StatusStopped` constant whose value was identical to `StatusPaused`, making it impossible to distinguish the two states
+* Duplicate compiled UUID regexes for tenant and snapshot IDs consolidated into a single shared `uuidRegex`
+* Removed unreachable validation guard in `OverwriteFromInstance` — the preceding empty-string check made the inner `if sourceInstanceID != ""` block impossible to reach
+* Compile error in `restoreFromSnapshot` example fixed — `nil` now passed correctly to `Snapshots.List` instead of an untyped string constant
+* `GetMetricValue` now returns a descriptive error rather than panicking when called with a nil metrics response
+* `test_helpers.go` and `test_helpers_types.go` converted to `_test.go` files — test mocks no longer compile into the production binary
+* `gDSSessionService.List` and `Get` now log cancelled-context errors consistently with all other service methods
+* `cmekService.List` tenant ID validation simplified — fragile length-based switch with incorrect error message replaced with a direct call to `ValidateTenantID`
+* `gDSSessionService.Create` and `Estimate` now return a clear error immediately when called with a nil request, rather than marshalling `null` to the API
+### Security
+* Added `CreateInstanceData.String()` to redact the `Password` field, preventing credentials leaking via logs or fmt.Printf
+* `WithBaseURL` now rejects non-HTTPS URLs to prevent OAuth tokens being transmitted in plaintext
+
+## v1.8.1 - 2026-04-23
+### Added
+*  `User-Agent` header now includes the library version (e.g. `aura-go-client/v1.8.0`)
+### Changed
+* Mock types moved from `internal/api` and `internal/httpClient` to `internal/testutil` — mocks no longer compile into production builds
+* URL construction now trims stray leading/trailing slashes to prevent double-slash in API endpoint paths
+### Fixed
+* Removed `StatusStopped` constant whose value was identical to `StatusPaused`, masking genuine "stopped" states
+* Added `StatusRestoring` to fix misspelled `StatusRestroying`; deprecated alias retained for backwards compatibility
+* Removed unreachable validation guard in `OverwriteFromInstance` — empty-string check above it already returned
+* `GetInstanceHealth` no longer stacks two independent timeouts; calls internal `doFetchRawMetrics` directly
+* `tenantService.List`, `Get`, and `GetMetrics` now check for a pre-cancelled context, matching all other services
+* Removed redundant HTTP 200-status check in `FetchRawMetrics` — the API layer already converts non-2xx to errors
+* Duplicate compiled UUID regexes for tenant and snapshot IDs consolidated into a single `uuidRegex`
+* `TruncateString` now slices by rune count instead of bytes, preventing truncation mid-multibyte character
+### Security
+* `CreateInstanceData.String()` added to redact the `Password` field — prevents credentials leaking via logs or fmt.Printf
+* WithBaseURL now rejects non-HTTPS URLs to protect OAuth tokens in transit
+
 ## v1.8.0 - 2026-03-24
 ### Added
 * Validation for instance and snapshort ID in snapshot service
